requirements.txt
docopt==0.6.2
invokust==0.52
locustio==0.11.0
requests==2.21.0
